refactor(raijin): split initializer ownership boundaries
