+++ yarn/raijin/src/cli.ts
@@ -1,4 +1,4 @@
-import { runRaijinInitializer } from './initializer.js'
+import { runRaijinInitializer } from './initializer/index.js'
 
 try {
   await runRaijinInitializer({ argv: process.argv.slice(2) })
