+++ yarn/raijin/src/initializer/exceptions/usage.exception.ts
@@ -1,4 +1,4 @@
-import { RAIJIN_INITIALIZER_USAGE_MESSAGE } from '../errors.js'
+import { RAIJIN_INITIALIZER_USAGE_MESSAGE } from '../../errors.js'
 
 export class RaijinInitializerUsageException extends Error {
   constructor() {
