+++ yarn/raijin/src/runtime/installer.interfaces.ts
@@ -1,4 +1,4 @@
-import type { FetchLike } from './runtime-download.interfaces.js'
+import type { FetchLike } from './download.interfaces.js'
 
 export interface InstallRaijinRuntimeOptions {
   cwd: string
