+++ yarn/raijin/src/yarn/exceptions/command.exception.ts
@@ -1,4 +1,4 @@
-import { createYarnCommandFailureMessage } from '../errors.js'
+import { createYarnCommandFailureMessage } from '../../errors.js'
 
 export class RaijinYarnCommandException extends Error {
   constructor(args: Array<string>) {
